+++ src/Interpolator.js
@@ -30,7 +30,7 @@ class Interpolator {
     this.logger = baseLogger.create('interpolator');
 
     this.options = options;
-    this.format = (options.interpolation && options.interpolation.format) || ((value) => value);
+    this.format = options?.interpolation?.format || ((value) => value);
     this.init(options);
   }
 
@@ -93,7 +93,7 @@ class Interpolator {
 
   resetRegExp() {
     const getOrResetRegExp = (existingRegExp, pattern) => {
-      if (existingRegExp && existingRegExp.source === pattern) {
+      if (existingRegExp?.source === pattern) {
         existingRegExp.lastIndex = 0;
         return existingRegExp;
       }
@@ -159,10 +159,10 @@ class Interpolator {
     this.resetRegExp();
 
     const missingInterpolationHandler =
-      (options && options.missingInterpolationHandler) || this.options.missingInterpolationHandler;
+      options?.missingInterpolationHandler || this.options.missingInterpolationHandler;
 
     const skipOnVariables =
-      options && options.interpolation && options.interpolation.skipOnVariables !== undefined
+      options?.interpolation?.skipOnVariables !== undefined
         ? options.interpolation.skipOnVariables
         : this.options.interpolation.skipOnVariables;
 
@@ -236,7 +236,7 @@ class Interpolator {
       const matchedSingleQuotes = optionsString.match(/'/g);
       const matchedDoubleQuotes = optionsString.match(/"/g);
       if (
-        (matchedSingleQuotes && matchedSingleQuotes.length % 2 === 0 && !matchedDoubleQuotes) ||
+        ((matchedSingleQuotes?.length ?? 0) % 2 === 0 && !matchedDoubleQuotes) ||
         matchedDoubleQuotes.length % 2 !== 0
       ) {
         optionsString = optionsString.replace(/'/g, '"');