+++ test/compatibility/v4/lib/utils.js
@@ -1,0 +1,3 @@
+export const isString = (obj) => typeof obj === 'string';
+
+export const getCleanedCode = (code) => code && code.replace('_', '-');