+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+## 24.0.0
+
+**This is a major breaking release:**
+
+- remove support for older environments
+- remove old i18next JSON formats
+  - To convert your existing v3 translations to the v4 format, have a look at [i18next-v4-format-converter](https://github.com/i18next/i18next-v4-format-converter) or this [web tool](https://i18next.github.io/i18next-v4-format-converter-web/).
+- remove support for compatibility to v1 API
+- Intl API is mandatory now and will not fallback anymore
+- possible compatibility layer for older formats: `test/compatibility/v4/v4Compatibility.js`
+- rename `initImmediate` to `initAsync`
+- fallback to `dev` language if plural rule not found
+- remove TypeScript v4 support. TypeScript v5 is now an optional peer dependency
+- addresses
+  - [2244](https://github.com/i18next/i18next/pull/2244)
+  - [2184](https://github.com/i18next/i18next/pull/2184)
+  - [2213](https://github.com/i18next/i18next/issues/2213)
+  - [2206](https://github.com/i18next/i18next/pull/2206)
+  - [2208](https://github.com/i18next/i18next/issues/2208)
+  - [2148](https://github.com/i18next/i18next/issues/2148)
+  - [2254](https://github.com/i18next/i18next/issues/2254)
+
+➡️ check out the [migration guide](https://www.i18next.com/misc/migration-guide#v23.x.x-to-v24.0.0)
+
 ## 23.16.8
 
 - fix(plural): Create cache entry for PluralRules created as part of the fallback flow [2256](https://github.com/i18next/i18next/pull/2256)